Refactor HunyuanImage3 shared image processing
