+++ vllm_omni/diffusion/models/hunyuan_image3/hunyuan_image3_transformer.py
@@ -76,6 +76,12 @@ from vllm_omni.diffusion.forward_context import set_forward_context_denoise_step
 from vllm_omni.diffusion.layers.norm import RMSNorm
 from vllm_omni.diffusion.layers.rope import RotaryEmbedding
 from vllm_omni.diffusion.models.hunyuan_image3.hunyuan_fused_moe import HunyuanFusedMoE
+from vllm_omni.diffusion.models.hunyuan_image3.image_processing import (
+    HUNYUAN_IMAGE3_EXTRA_RESOLUTIONS,
+    Resolution,
+    ResolutionGroup,
+    resize_and_crop,
+)
 from vllm_omni.model_executor.layers.timestep_embedding import timestep_embedding
 
 logger = logging.getLogger(__name__)
@@ -442,140 +448,6 @@ def repeat_kv(hidden_states: torch.Tensor, n_rep: int) -> torch.Tensor:
 
 def default(value, default_value):
     return value if value is not None else default_value
-
-
-class Resolution:
-    def __init__(self, size, *args):
-        if isinstance(size, str):
-            if "x" in size:
-                size = size.split("x")
-                size = (int(size[0]), int(size[1]))
-            else:
-                size = int(size)
-        if len(args) > 0:
-            size = (size, args[0])
-        if isinstance(size, int):
-            size = (size, size)
-
-        self.h = self.height = size[0]
-        self.w = self.width = size[1]
-        self.r = self.ratio = self.height / self.width
-
-    def __getitem__(self, idx):
-        if idx == 0:
-            return self.h
-        elif idx == 1:
-            return self.w
-        else:
-            raise IndexError(f"Index {idx} out of range")
-
-    def __str__(self):
-        return f"{self.h}x{self.w}"
-
-
-# Baked-in extras matching the official model's
-# `HunyuanImage3ImageProcessor.vae_reso_group` (image_processor.py:147-152).
-# These four aspect buckets sit at ratio_token indices 33-36 in the trained
-# model and the AR was trained to address them, so any deviation breaks the
-# ratio-token vocab → output-shape lookup.
-HUNYUAN_IMAGE3_EXTRA_RESOLUTIONS: tuple[str, ...] = (
-    "1024x768",
-    "1280x720",
-    "768x1024",
-    "720x1280",
-)
-
-
-class ResolutionGroup:
-    def __init__(self, base_size=None, step=None, align=1, extra_resolutions=None):
-        self.align = align
-        self.base_size = base_size
-        assert base_size % align == 0, f"base_size {base_size} is not divisible by align {align}"
-        if base_size is not None and not isinstance(base_size, int):
-            raise ValueError(f"base_size must be None or int, but got {type(base_size)}")
-        if step is None:
-            step = base_size // 16
-        if step is not None and step > base_size // 2:
-            raise ValueError(f"step must be smaller than base_size // 2, but got {step} > {base_size // 2}")
-
-        self.step = step
-        self.data = self._calc_by_step()
-
-        if extra_resolutions is not None:
-            for er in extra_resolutions:
-                if not any(r.ratio == er.ratio for r in self.data):
-                    self.data.append(er)
-
-        self.ratio = np.array([x.ratio for x in self.data])
-        self.attr = ["" for _ in range(len(self.data))]
-        self.prefix_space = 0
-
-    def __len__(self):
-        return len(self.data)
-
-    def __getitem__(self, idx):
-        return self.data[idx]
-
-    def __repr__(self):
-        prefix = self.prefix_space * " "
-        prefix_close = (self.prefix_space - 4) * " "
-        res_str = f"ResolutionGroup(base_size={self.base_size}, step={self.step}, data="
-        attr_maxlen = max([len(x) for x in self.attr] + [5])
-        res_str += (
-            f"\n{prefix}ID: height width   ratio {' ' * max(0, attr_maxlen - 4)}count  h/16 w/16    tokens\n{prefix}"
-        )
-        res_str += ("\n" + prefix).join(
-            [
-                f"{i:2d}: ({x.h:4d}, {x.w:4d})  {self.ratio[i]:.4f}  {self.attr[i]:>{attr_maxlen}s}  "
-                f"({x.h // 16:3d}, {x.w // 16:3d})  {x.h // 16 * x.w // 16:6d}"
-                for i, x in enumerate(self.data)
-            ]
-        )
-        res_str += f"\n{prefix_close})"
-        return res_str
-
-    def _calc_by_step(self):
-        assert self.align <= self.step, f"align {self.align} must be smaller than step {self.step}"
-
-        min_height = self.base_size // 2
-        min_width = self.base_size // 2
-        max_height = self.base_size * 2
-        max_width = self.base_size * 2
-
-        resolutions = [Resolution(self.base_size, self.base_size)]
-
-        cur_height, cur_width = self.base_size, self.base_size
-        while True:
-            if cur_height >= max_height and cur_width <= min_width:
-                break
-
-            cur_height = min(cur_height + self.step, max_height)
-            cur_width = max(cur_width - self.step, min_width)
-            resolutions.append(Resolution(cur_height // self.align * self.align, cur_width // self.align * self.align))
-
-        cur_height, cur_width = self.base_size, self.base_size
-        while True:
-            if cur_height <= min_height and cur_width >= max_width:
-                break
-
-            cur_height = max(cur_height - self.step, min_height)
-            cur_width = min(cur_width + self.step, max_width)
-            resolutions.append(Resolution(cur_height // self.align * self.align, cur_width // self.align * self.align))
-
-        resolutions = sorted(resolutions, key=lambda x: x.ratio)
-
-        return resolutions
-
-    def get_target_size(self, width, height):
-        ratio = height / width
-        idx = np.argmin(np.abs(self.ratio - ratio))
-        reso = self.data[idx]
-        return reso.w, reso.h
-
-    def get_base_size_and_ratio_index(self, width, height):
-        ratio = height / width
-        idx = np.argmin(np.abs(self.ratio - ratio))
-        return self.base_size, idx
 
 
 class ImageInfo:
@@ -1471,25 +1343,7 @@ class HunyuanImage3ImageProcessor:
         target_size: tuple[int, int],
         crop_type: str = "center",
     ) -> Image.Image:
-        tw, th = target_size
-        if crop_type == "resize":
-            # `infer_align_image_size=True` follows HF's direct-resize path
-            # for condition images; the generated output is aligned back to
-            # the input aspect ratio in `postprocess_outputs`.
-            return image.resize((tw, th), resample=Image.Resampling.LANCZOS)
-        w, h = image.size
-        tr = th / tw
-        r = h / w
-        if r < tr:
-            resize_height = th
-            resize_width = int(round(th / h * w))
-        else:
-            resize_width = tw
-            resize_height = int(round(tw / w * h))
-        image = image.resize((resize_width, resize_height), resample=Image.Resampling.LANCZOS)
-        crop_top = int(round((resize_height - th) / 2.0))
-        crop_left = int(round((resize_width - tw) / 2.0))
-        return image.crop((crop_left, crop_top, crop_left + tw, crop_top + th))
+        return resize_and_crop(image, target_size, crop_type=crop_type)
 
     def postprocess_outputs(
         self,
